requirements.txt
Pillow==9.4.0
opencv-python-headless==4.7.0.72
numpy==1.24.2
python-dotenv==1.0.0
Flask==2.2.3
pandas==1.5.3
onnxruntime==1.14.1
gunicorn==20.1.0